Rename package for automata with predicates
The finite state module aims to have various kind of finite state
automata and transducers, each of which will live in a package
describing what kind of machine it is. The `automaton` name is too
generic and can be confusing in the presence of several implementations.
